move translations
