+++ package.json
@@ -26,7 +26,7 @@
         "cypress:run": "cypress run",
         "gql": "graphql-codegen --config codegen.yml",
         "i18n": "i18next",
-        "i18n:regacy": "extract-messages -d en -l=en,ja -o src/translations/legacy -f yaml 'src/**/*.ts*' --extractFromFormatMessageCall",
+        "i18n:legacy": "extract-messages -d en -l=en,ja -o src/i18n/translations/legacy -f yaml 'src/**/*.ts*' --extractFromFormatMessageCall",
         "gen:doc:plugin": "ts-node -O '{\"module\":\"CommonJS\"}' ./bin/pluginDoc"
     },
     "engines": {
