+++ src/i18n/i18n.ts
@@ -2,8 +2,8 @@ import i18n from "i18next";
 import LanguageDetector from "i18next-browser-languagedetector";
 import { initReactI18next } from "react-i18next";
 
-import en from "../translations/en.yml";
-import ja from "../translations/ja.yml";
+import en from "./translations/en.yml";
+import ja from "./translations/ja.yml";
 
 const resources = {
   en: {
