+++ src/i18n/translations/legacy/ja.yml
